🚚 Move all Tracker code to correct package
The Jetpack Navigation Compose does not have support for Dynamic Feature
Modules at the moment. At first, I added all the code inside app module
temporarily. But know I thought that I can use another Activity in the
DFM and open it by deep link while there is no native support in
Compose.
