+++ features/tracker/build.gradle
@@ -4,4 +4,7 @@ dependencies {
     implementation project(":domain")
 
     implementation Deps.koin.android
+    implementation Deps.coroutines.core
+
+    testImplementation project(':libraries:test')
 }
