+++ features/tracker/src/main/java/com/escodro/tracker/presentation/TrackerViewModel.kt
@@ -1,8 +1,8 @@
-package com.escodro.alkaa.presentation.tracker.presentation
+package com.escodro.tracker.presentation
 
 import androidx.lifecycle.ViewModel
-import com.escodro.alkaa.presentation.tracker.mapper.TrackerMapper
 import com.escodro.domain.usecase.tracker.LoadCompletedTasksByPeriod
+import com.escodro.tracker.mapper.TrackerMapper
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.collect
